requirements.txt
brotlipy==0.7.0
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
cryptography==3.4.7
idna==2.7
mkl-fft==1.3.0
mkl-random==1.1.0
mkl-service==2.3.0
msedge-selenium-tools==3.141.3
numpy==1.19.2
Pillow==8.3.1
pip==21.1.3
pycparser==2.20
pyOpenSSL==20.0.1
PySocks==1.7.1
PyYAML==5.4.1
requests==2.25.1
selenium==3.141.0
setuptools==52.0.0.post20210125
six==1.16.0
torch==1.9.0
torchvision==0.10.0
typing-extensions==3.10.0.0
urllib3==1.24.3
wheel==0.36.2
win-inet-pton==1.1.0
wincertstore==0.2
